+++ htpclient/initialize.py
@@ -210,18 +210,17 @@ class Initialize:
             self.__check_url(args)
         else:
             logging.debug("Connection test successful!")
-
+            
         if args.cpu_only is not None and args.cpu_only:
             logging.debug("Setting agent to be CPU only..")
             self.config.set_value('cpu-only', True)
 
-    @staticmethod
-    def __build_directories():
-        if not os.path.isdir("crackers"):
-            os.mkdir("crackers")
-        if not os.path.isdir("files"):
-            os.mkdir("files")
-        if not os.path.isdir("hashlists"):
-            os.mkdir("hashlists")
-        if not os.path.isdir("preprocessor"):
-            os.mkdir("preprocessor")
+    def __build_directories(self):
+        if not os.path.isdir(self.config.get_value('crackers-path')):
+            os.makedirs(self.config.get_value('crackers-path'))
+        if not os.path.isdir(self.config.get_value('files-path')):
+            os.makedirs(self.config.get_value('files-path'))
+        if not os.path.isdir(self.config.get_value('hashlists-path')):
+            os.makedirs(self.config.get_value('hashlists-path'))
+        if not os.path.isdir(self.config.get_value('preprocessors-path')):
+            os.makedirs(self.config.get_value('preprocessors-path'))